Extract src/utils/rtlUtils.
